requirements.txt
Django==1.5.1
South==0.7.6
argparse==1.2.1
boto==2.9.2
distribute==0.6.24
dj-database-url==0.2.1
django-grappelli==2.4.4
django-haystack==2.0.0
django-tastypie==0.9.15
gunicorn==0.17.4
mimeparse==0.1.3
psycopg2==2.5
pyelasticsearch==0.5
python-dateutil==2.1
requests==1.2.3
simplejson==3.3.0
six==1.3.0
wsgiref==0.1.2
